requirements.txt
advanced_alchemy==0.8.3
aiosqlite==0.20.0
greenlet== 3.0.3
jsonschema==4.20.0
litestar== 2.7.1
loguru==0.7.2
pydantic== 2.5.2
SQLAlchemy==2.0.29
typing_extensions==4.9.0
ulid-py==1.1.0
urllib3==2.1.0
uvicorn==0.24.0.post1